- Version up to 0.98
- Fixed TranslateName() to support FAT32 variation of WinNT family.

- Fixed FS_PATHINFO to support level 7 correctly.

- Both /FS and /FL option is not supported any more.

- Added 'Directory Circularity Check' for FS_COPY and FS_MOVE
  in case of mixed path of short and long name.

- Added /CP option to CACHEF32.EXE to specify codepage in paramter.

+++ src/ufat32.c
@@ -1066,7 +1066,7 @@ USHORT rc;
                 if (pCD->fFix)
                 {
                      strcpy(Mark.szFileName, pbPath);
-                     Mark.fEAS = pDir->fEAS == FILE_HAS_OLD_EAS ? FILE_HAS_EAS : FILE_HAS_CRITICAL_EAS;
+                     Mark.fEAS = ( BYTE )( pDir->fEAS == FILE_HAS_OLD_EAS ? FILE_HAS_EAS : FILE_HAS_CRITICAL_EAS );
                      rc = DosDevIOCtl2(NULL, 0,
                                        (PVOID)&Mark, sizeof Mark,
                                        FAT32_SETEAS, IOCTL_FAT32, pCD->hDisk);
